Revert "Fix toStartOfInterval with negative values (toStartOfIntevalAllowNegative) "
