+++ docs/en/sql-reference/functions/date-time-functions.md
@@ -1642,20 +1642,6 @@ Result:
 **See Also**
 - [date_trunc](#date_trunc)
 
-**Note:** in case of using negative datetime values (before `1970-01-01`) as arguments for this function, please consider using the [`toStartOfIntervalAllowNegative`](#toStartOfIntervalAllowNegative) function, which supports these cases.
-
-## toStartOfIntervalAllowNegative
-
-Works just like the `toStartOfInterval` function but works correctly with negative date/datetime values (before 1970-01-01) as function arguments.
-
-**Syntax**
-
-```sql
-toStartOfIntervalAllowNegative(value, INTERVAL x unit[, time_zone])
-toStartOfIntervalAllowNegative(value, INTERVAL x unit[, origin[, time_zone]])
-```
-Aliases: `time_bucket_allow_negative`, `date_bin_allow_negative`.
-
 ## toTime
 
 Converts a date with time to a certain fixed date, while preserving the time.
