--- tests/queries/0_stateless/03318_toStartOfIntervalAllowNegative.reference
@@ -1,24 +1,0 @@
---- Negative DateTime64
-1960-01-01
-1960-01-01
-1960-03-01
-1960-02-29
-1
-1960-03-03 00:00:00
-1960-03-03 12:00:00
-1960-03-03 12:55:00
-1960-03-03 12:55:56
-1960-03-03 12:55:55.123
---- Negative Date32
-1960-01-01
-1960-01-01
-1960-03-01
-1960-02-29
-1
-1960-03-03 00:00:00
---- Non-const arguments
-1970-03-02
-1960-02-29
-1970-03-02
-1960-02-29
-1970-03-02
