+++ utils/check-style/aspell-ignore/en/aspell-dict.txt
@@ -2980,7 +2980,6 @@ toStartOfFiveMinutes
 toStartOfHour
 toStartOfISOYear
 toStartOfInterval
-toStartOfIntervalAllowNegative
 toStartOfMicrosecond
 toStartOfMillisecond
 toStartOfMinute
